Initial commit
This is a minimal feature complete vector/matrix library.  This provides
vector sand matrices that follow all of the correct mathematical
behavior.  It may not be the most computationally efficient, but it
works.  On the To Do list is making the inner products and mathematical
operations more efficient by using LAPACK.